Rename main_widget.py to manager.py
This is in preparation to make the plugin a SpyderPluginV2 one.
