+++ spyder_env_manager/spyder/widgets/manager.py
